// lib/types.ts
// Tipos compartidos para toda la aplicación

export type LegParams = {
  footLength: number
  archHeight: number
  heelRadius: number
  toeCount: number
  tibiaLength: number
  femurLength: number
  legThickness: number
  kneeAngle: number
  ankleAngle: number
  hipAngle: number
  footRotation: number
  stepAngle: number
  verticalShift: number
  springStiffness: number
  dampingFactor: number
}